:heavy_plus_sign: Add Records class
